upd gnl

+++ includes/ft_io.h
@@ -6,12 +6,13 @@
 /*   By: svet <svet@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/18 15:26:49 by svet              #+#    #+#             */
-/*   Updated: 2020/05/26 14:13:32 by svet             ###   ########.fr       */
+/*   Updated: 2020/10/19 12:55:33 by svet             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_IO_H
 # define FT_IO_H
+# include <sys/_types/_ssize_t.h>
 
 void	ft_putchar_fd(int c, int fd);
 void	ft_putstr_fd(char const *s, int fd);
@@ -25,6 +26,6 @@ void	ft_putnbr(int n);
 void	ft_putlnbr_base_fd(long n, int base, int isupper, int fd);
 void	ft_putlnbr_base(long n, int base, int isupper);
 
-int		ft_getline(const int fd, char **line);
+ssize_t	ft_getline(const int fd, char **line);
 
 #endif
